fix(floor-plan): fix API prefix nesting and response shape, seed test data
- Fix double /api nesting: remove /api from mod.rs nests since app.rs already handles it
- Fix response shape: backend returns {zones} instead of {data} to match frontend
- Unwrap {data} wrapper from get_table/update_status/reserve endpoints
- Fix FLOAT4/FLOAT8 type mismatch in migration (REAL -> DOUBLE PRECISION)
- Run migrations and seed 16 tables across 3 zones for testing

+++ backend/migrations/branch/20250101000002_tables.sql
@@ -8,8 +8,8 @@ CREATE TABLE tables (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     name VARCHAR(100) NOT NULL,
     zone_id UUID NOT NULL REFERENCES table_zones(id),
-    x REAL NOT NULL DEFAULT 0,
-    y REAL NOT NULL DEFAULT 0,
+    x DOUBLE PRECISION NOT NULL DEFAULT 0,
+    y DOUBLE PRECISION NOT NULL DEFAULT 0,
     seats INT NOT NULL DEFAULT 4,
     status VARCHAR(20) NOT NULL DEFAULT 'available'
         CHECK (status IN ('available', 'occupied', 'dirty', 'reserved')),
